refactor: split identity into embedding/rerank/engram, add zero-config embedding
- Delete identity/Evolution*/resolveEvolution from config schema
- Add top-level embedding, rerank, engram fields to Config.Info
- Split config.ts (9460→972 lines) into config/schema.ts (1285 lines)
- Add config migration: identity.* → new format in user config files
- Rewrite embedding.ts: local Transformers.js fallback + warmup
- Add BM25 fallback in memory-recall.ts when embeddings unavailable
- Make rerank opt-in (only when config.rerank.apiKey configured)
- Replace passive/active naming with memory/experience
- Delete cli/cmd/identity.ts, create cli/cmd/engram.ts
- Update setup validation, runtime reload, all 15+ consumer files
- All 156 tests pass, typecheck zero errors
