+++ packages/synergy/package.json
@@ -72,6 +72,7 @@
     "@ericsanchezok/synergy-sdk": "workspace:*",
     "@ericsanchezok/synergy-util": "workspace:*",
     "@hono/zod-validator": "catalog:",
+    "@huggingface/transformers": "4.2.0",
     "@larksuiteoapi/node-sdk": "1.58.0",
     "@modelcontextprotocol/sdk": "1.25.2",
     "@openrouter/ai-sdk-provider": "1.5.2",
