+++ packages/synergy/src/config/config.ts
@@ -30,6 +30,7 @@ import { existsSync } from "fs"
 import { ConfigSet } from "./set"
 import { loadFragments } from "./fragment"
 import { RuntimeSchema } from "../runtime/schema"
+import * as Schema from "./schema"
 
 export namespace Config {
   const log = Log.create({ service: "config" })
@@ -39,6 +40,76 @@ export namespace Config {
     insertSpaces: true,
     eol: "\n",
   } as const
+
+  // ── Schema re-exports from ./schema.ts ──
+
+  export const McpLocal = Schema.McpLocal
+  export const McpOAuth = Schema.McpOAuth
+  export type McpOAuth = Schema.McpOAuth
+  export const McpRemote = Schema.McpRemote
+  export const Mcp = Schema.Mcp
+  export type Mcp = Schema.Mcp
+  export const FeishuGroupSessionScope = Schema.FeishuGroupSessionScope
+  export type FeishuGroupSessionScope = Schema.FeishuGroupSessionScope
+  export const ChannelFeishuAccount = Schema.ChannelFeishuAccount
+  export type ChannelFeishuAccount = Schema.ChannelFeishuAccount
+  export const ChannelFeishu = Schema.ChannelFeishu
+  export type ChannelFeishu = Schema.ChannelFeishu
+  export const Holos = Schema.Holos
+  export type Holos = Schema.Holos
+  export const Channel = Schema.Channel
+  export type Channel = Schema.Channel
+  export const EmailSmtp = Schema.EmailSmtp
+  export type EmailSmtp = Schema.EmailSmtp
+  export const EmailImap = Schema.EmailImap
+  export type EmailImap = Schema.EmailImap
+  export const EmailFrom = Schema.EmailFrom
+  export type EmailFrom = Schema.EmailFrom
+  export const Email = Schema.Email
+  export type Email = Schema.Email
+  export const PermissionAction = Schema.PermissionAction
+  export type PermissionAction = Schema.PermissionAction
+  export const PermissionObject = Schema.PermissionObject
+  export type PermissionObject = Schema.PermissionObject
+  export const PermissionRule = Schema.PermissionRule
+  export type PermissionRule = Schema.PermissionRule
+  export const Permission = Schema.Permission
+  export type Permission = Schema.Permission
+  export const Command = Schema.Command
+  export type Command = Schema.Command
+  export const Agent = Schema.Agent
+  export type Agent = Schema.Agent
+  export const ExternalAgentConfig = Schema.ExternalAgentConfig
+  export type ExternalAgentConfig = Schema.ExternalAgentConfig
+  export const Keybinds = Schema.Keybinds
+  export const Server = Schema.Server
+  export const CategoryConfig = Schema.CategoryConfig
+  export type CategoryConfig = Schema.CategoryConfig
+  export const Layout = Schema.Layout
+  export type Layout = Schema.Layout
+  export const Learning = Schema.Learning
+  export type Learning = Schema.Learning
+  export const PassiveRetrieval = Schema.PassiveRetrieval
+  export type PassiveRetrieval = Schema.PassiveRetrieval
+  export const REWARD_WEIGHT_DEFAULTS = Schema.REWARD_WEIGHT_DEFAULTS
+  export const LEARNING_DEFAULTS = Schema.LEARNING_DEFAULTS
+  export const PASSIVE_RETRIEVAL_DEFAULTS = Schema.PASSIVE_RETRIEVAL_DEFAULTS
+  export const MEMORY_CATEGORIES = Schema.MEMORY_CATEGORIES
+  export type MemoryCategory = Schema.MemoryCategory
+  export const EmbeddingConfig = Schema.EmbeddingConfig
+  export type EmbeddingConfig = Schema.EmbeddingConfig
+  export const RerankConfig = Schema.RerankConfig
+  export type RerankConfig = Schema.RerankConfig
+  export const MemoryConfig = Schema.MemoryConfig
+  export type MemoryConfig = Schema.MemoryConfig
+  export const ExperienceConfig = Schema.ExperienceConfig
+  export type ExperienceConfig = Schema.ExperienceConfig
+  export const EngramConfig = Schema.EngramConfig
+  export type EngramConfig = Schema.EngramConfig
+  export const Provider = Schema.Provider
+  export type Provider = Schema.Provider
+  export const Info = Schema.Info
+  export type Info = Schema.Info
 
   // Custom merge function that concatenates array fields instead of replacing them
   function mergeConfigConcatArrays(target: Info, source: Info): Info {
@@ -204,9 +275,18 @@ export namespace Config {
       if (result.compaction.overflowThreshold === undefined) result.compaction.overflowThreshold = 0.85
       if (result.compaction.maxHistoryImages === undefined) result.compaction.maxHistoryImages = 8
     }
-    if (result.identity) {
-      if (result.identity.evolution === undefined) result.identity.evolution = true
-      if (result.identity.autonomy === undefined) result.identity.autonomy = true
+    if (result.engram) {
+      if (result.engram.memory === undefined) result.engram.memory = { enabled: true }
+      if (result.engram.memory && !result.engram.memory.retrieval) {
+        result.engram.memory.retrieval = { simThreshold: 0.7, topK: 3 }
+      }
+      if (result.engram.memory && !result.engram.memory.dedup) {
+        result.engram.memory.dedup = { threshold: 0.75 }
+      }
+      if (result.engram.experience === undefined) {
+        result.engram.experience = { encode: true, retrieve: true, learning: { ...LEARNING_DEFAULTS } }
+      }
+      if (result.engram.autonomy === undefined) result.engram.autonomy = true
     }
 
     if (!result.username) result.username = os.userInfo().username
@@ -352,1310 +432,6 @@ export namespace Config {
     }
     return plugins
   }
-
-  export const McpLocal = z
-    .object({
-      type: z.literal("local").describe("Type of MCP server connection"),
-      command: z.string().array().describe("Command and arguments to run the MCP server"),
-      environment: z
-        .record(z.string(), z.string())
-        .optional()
-        .describe("Environment variables to set when running the MCP server"),
-      enabled: z.boolean().optional().describe("Enable or disable the MCP server on startup"),
-      timeout: z
-        .number()
-        .int()
-        .positive()
-        .optional()
-        .describe(
-          "Timeout in ms for fetching tools from the MCP server. Defaults to 5000 (5 seconds) if not specified.",
-        ),
-    })
-    .strict()
-    .meta({
-      ref: "McpLocalConfig",
-    })
-
-  export const McpOAuth = z
-    .object({
-      clientId: z
-        .string()
-        .optional()
-        .describe("OAuth client ID. If not provided, dynamic client registration (RFC 7591) will be attempted."),
-      clientSecret: z.string().optional().describe("OAuth client secret (if required by the authorization server)"),
-      scope: z.string().optional().describe("OAuth scopes to request during authorization"),
-    })
-    .strict()
-    .meta({
-      ref: "McpOAuthConfig",
-    })
-  export type McpOAuth = z.infer<typeof McpOAuth>
-
-  export const McpRemote = z
-    .object({
-      type: z.literal("remote").describe("Type of MCP server connection"),
-      url: z.string().describe("URL of the remote MCP server"),
-      enabled: z.boolean().optional().describe("Enable or disable the MCP server on startup"),
-      headers: z.record(z.string(), z.string()).optional().describe("Headers to send with the request"),
-      oauth: z
-        .union([McpOAuth, z.literal(false)])
-        .optional()
-        .describe(
-          "OAuth authentication configuration for the MCP server. Set to false to disable OAuth auto-detection.",
-        ),
-      timeout: z
-        .number()
-        .int()
-        .positive()
-        .optional()
-        .describe(
-          "Timeout in ms for fetching tools from the MCP server. Defaults to 5000 (5 seconds) if not specified.",
-        ),
-    })
-    .strict()
-    .meta({
-      ref: "McpRemoteConfig",
-    })
-
-  export const Mcp = z.discriminatedUnion("type", [McpLocal, McpRemote])
-  export type Mcp = z.infer<typeof Mcp>
-
-  export const FeishuGroupSessionScope = z
-    .enum(["group", "group_sender", "group_topic", "group_topic_sender"])
-    .describe(
-      "How group chat sessions are scoped: group = shared, group_sender = per sender, group_topic = per thread/topic, group_topic_sender = per thread+sender",
-    )
-  export type FeishuGroupSessionScope = z.infer<typeof FeishuGroupSessionScope>
-
-  export const ChannelFeishuAccount = z
-    .object({
-      enabled: z.boolean().optional().default(true),
-      appId: z.string().describe("Feishu app ID"),
-      appSecret: z.string().describe("Feishu app secret"),
-      domain: z
-        .enum(["feishu", "lark"])
-        .optional()
-        .describe("Feishu domain (feishu for China, lark for international)"),
-      allowDM: z.boolean().optional().default(true).describe("Allow direct messages"),
-      allowGroup: z.boolean().optional().default(true).describe("Allow group messages"),
-      requireMention: z.boolean().optional().default(true).describe("Require @mention in group chats"),
-      botOpenId: z.string().optional().describe("Bot open_id used to verify real @mentions in group chats"),
-      streaming: z.boolean().optional().default(true).describe("Enable streaming card updates"),
-      streamingThrottleMs: z
-        .number()
-        .int()
-        .positive()
-        .optional()
-        .default(100)
-        .describe("Minimum interval between streaming card updates in ms"),
-      groupSessionScope: FeishuGroupSessionScope.optional()
-        .default("group")
-        .describe("Session scoping strategy for group chats"),
-      inboundDebounceMs: z
-        .number()
-        .int()
-        .nonnegative()
-        .optional()
-        .default(0)
-        .describe("Debounce rapid-fire messages from the same sender in the same chat (0 = disabled)"),
-      resolveSenderNames: z
-        .boolean()
-        .optional()
-        .default(true)
-        .describe("Resolve sender display names via Feishu contact API"),
-      replyInThread: z.boolean().optional().default(false).describe("Reply in thread when message is part of a topic"),
-    })
-    .strict()
-    .meta({ ref: "ChannelFeishuAccountConfig" })
-  export type ChannelFeishuAccount = z.infer<typeof ChannelFeishuAccount>
-
-  export const ChannelFeishu = z
-    .object({
-      type: z.literal("feishu"),
-      accounts: z.record(z.string(), ChannelFeishuAccount),
-      domain: z.enum(["feishu", "lark"]).optional().describe("Default domain for all accounts"),
-      streaming: z.boolean().optional().default(true).describe("Default streaming setting for all accounts"),
-    })
-    .strict()
-    .meta({ ref: "ChannelFeishuConfig" })
-  export type ChannelFeishu = z.infer<typeof ChannelFeishu>
-
-  export const Holos = z
-    .object({
-      enabled: z.boolean().optional().default(true).describe("Enable the Holos runtime connection"),
-      apiUrl: z.string().optional().default("https://api.holosai.io").describe("Holos API base URL"),
-      wsUrl: z.string().optional().default("wss://api.holosai.io").describe("Holos WebSocket base URL"),
-      portalUrl: z
-        .string()
-        .optional()
-        .default("https://www.holosai.io")
-        .describe("Holos portal URL for browser-facing pages (bind/start)"),
-    })
-    .strict()
-    .meta({ ref: "HolosConfig" })
-  export type Holos = z.infer<typeof Holos>
-
-  export const Channel = z.discriminatedUnion("type", [ChannelFeishu])
-  export type Channel = z.infer<typeof Channel>
-
-  export const EmailSmtp = z
-    .object({
-      host: z.string().optional().describe("SMTP server hostname"),
-      port: z.number().int().positive().optional().describe("SMTP server port"),
-      secure: z.boolean().optional().describe("Use TLS/SSL for the SMTP connection"),
-      username: z.string().optional().describe("SMTP username"),
-      password: z.string().optional().describe("SMTP password or app token"),
-    })
-    .strict()
-    .meta({ ref: "EmailSmtpConfig" })
-  export type EmailSmtp = z.infer<typeof EmailSmtp>
-
-  export const EmailImap = z
-    .object({
-      host: z.string().optional().describe("IMAP server hostname"),
-      port: z.number().int().positive().optional().describe("IMAP server port"),
-      secure: z.boolean().optional().describe("Use TLS/SSL for the IMAP connection"),
-      username: z.string().optional().describe("IMAP username"),
-      password: z.string().optional().describe("IMAP password or app token"),
-    })
-    .strict()
-    .meta({ ref: "EmailImapConfig" })
-  export type EmailImap = z.infer<typeof EmailImap>
-
-  export const EmailFrom = z
-    .object({
-      address: z.string().optional().describe("Sender email address"),
-      name: z.string().optional().describe("Sender display name"),
-    })
-    .strict()
-    .meta({ ref: "EmailFromConfig" })
-  export type EmailFrom = z.infer<typeof EmailFrom>
-
-  export const Email = z
-    .object({
-      enabled: z.boolean().optional().describe("Enable email features"),
-      from: EmailFrom.optional().describe("Sender identity for outgoing emails"),
-      smtp: EmailSmtp.optional().describe("SMTP transport settings for outgoing emails"),
-      imap: EmailImap.optional().describe("IMAP settings for reading emails"),
-    })
-    .strict()
-    .meta({ ref: "EmailConfig" })
-  export type Email = z.infer<typeof Email>
-
-  export const PermissionAction = z.enum(["ask", "allow", "deny"]).meta({
-    ref: "PermissionActionConfig",
-  })
-  export type PermissionAction = z.infer<typeof PermissionAction>
-
-  export const PermissionObject = z.record(z.string(), PermissionAction).meta({
-    ref: "PermissionObjectConfig",
-  })
-  export type PermissionObject = z.infer<typeof PermissionObject>
-
-  export const PermissionRule = z.union([PermissionAction, PermissionObject]).meta({
-    ref: "PermissionRuleConfig",
-  })
-  export type PermissionRule = z.infer<typeof PermissionRule>
-
-  // Capture original key order before zod reorders, then rebuild in original order
-  const permissionPreprocess = (val: unknown) => {
-    if (typeof val === "object" && val !== null && !Array.isArray(val)) {
-      return { __originalKeys: Object.keys(val), ...val }
-    }
-    return val
-  }
-
-  const permissionTransform = (x: unknown): Record<string, PermissionRule> => {
-    if (typeof x === "string") return { "*": x as PermissionAction }
-    const obj = x as { __originalKeys?: string[] } & Record<string, unknown>
-    const { __originalKeys, ...rest } = obj
-    if (!__originalKeys) return rest as Record<string, PermissionRule>
-    const result: Record<string, PermissionRule> = {}
-    for (const key of __originalKeys) {
-      if (key in rest) result[key] = rest[key] as PermissionRule
-    }
-    return result
-  }
-
-  export const Permission = z
-    .preprocess(
-      permissionPreprocess,
-      z
-        .object({
-          __originalKeys: z.string().array().optional(),
-          read: PermissionRule.optional(),
-          edit: PermissionRule.optional(),
-          glob: PermissionRule.optional(),
-          grep: PermissionRule.optional(),
-          list: PermissionRule.optional(),
-          bash: PermissionRule.optional(),
-          task: PermissionRule.optional(),
-          external_directory: PermissionRule.optional(),
-          todowrite: PermissionAction.optional(),
-          todoread: PermissionAction.optional(),
-          dagwrite: PermissionAction.optional(),
-          dagread: PermissionAction.optional(),
-          question: PermissionAction.optional(),
-          webfetch: PermissionAction.optional(),
-          websearch: PermissionAction.optional(),
-          download: PermissionAction.optional(),
-          lsp: PermissionRule.optional(),
-          doom_loop: PermissionAction.optional(),
-        })
-        .catchall(PermissionRule)
-        .or(PermissionAction),
-    )
-    .transform(permissionTransform)
-    .meta({
-      ref: "PermissionConfig",
-    })
-  export type Permission = z.infer<typeof Permission>
-
-  export const Command = z.object({
-    template: z.string(),
-    description: z.string().optional(),
-    agent: z.string().optional(),
-    model: z.string().optional(),
-  })
-  export type Command = z.infer<typeof Command>
-
-  export const Agent = z
-    .object({
-      model: z.string().optional(),
-      temperature: z.number().optional(),
-      top_p: z.number().optional(),
-      prompt: z.string().optional(),
-      tools: z.record(z.string(), z.boolean()).optional().describe("@deprecated Use 'permission' field instead"),
-      disable: z.boolean().optional(),
-      description: z.string().optional().describe("Description of when to use the agent"),
-      mode: z.enum(["subagent", "primary", "all"]).optional(),
-      hidden: z
-        .boolean()
-        .optional()
-        .describe("Hide this subagent from the @ autocomplete menu (default: false, only applies to mode: subagent)"),
-      options: z.record(z.string(), z.any()).optional(),
-      color: z
-        .string()
-        .regex(/^#[0-9a-fA-F]{6}$/, "Invalid hex color format")
-        .optional()
-        .describe("Hex color code for the agent (e.g., #FF5733)"),
-      steps: z
-        .number()
-        .int()
-        .positive()
-        .optional()
-        .describe("Maximum number of agentic iterations before forcing text-only response"),
-      maxSteps: z.number().int().positive().optional().describe("@deprecated Use 'steps' field instead."),
-      permission: Permission.optional(),
-    })
-    .catchall(z.any())
-    .transform((agent, ctx) => {
-      const knownKeys = new Set([
-        "name",
-        "model",
-        "prompt",
-        "description",
-        "temperature",
-        "top_p",
-        "mode",
-        "hidden",
-        "color",
-        "steps",
-        "maxSteps",
-        "options",
-        "permission",
-        "disable",
-        "tools",
-      ])
-
-      // Extract unknown properties into options
-      const options: Record<string, unknown> = { ...agent.options }
-      for (const [key, value] of Object.entries(agent)) {
-        if (!knownKeys.has(key)) options[key] = value
-      }
-
-      // Convert legacy tools config to permissions
-      const permission: Permission = {}
-      for (const [tool, enabled] of Object.entries(agent.tools ?? {})) {
-        const action = enabled ? "allow" : "deny"
-        // write, edit, patch, multiedit all map to edit permission
-        if (tool === "write" || tool === "edit" || tool === "patch" || tool === "multiedit") {
-          permission.edit = action
-        } else {
-          permission[tool] = action
-        }
-      }
-      Object.assign(permission, agent.permission)
-
-      // Convert legacy maxSteps to steps
-      const steps = agent.steps ?? agent.maxSteps
-
-      return { ...agent, options, permission, steps } as typeof agent & {
-        options?: Record<string, unknown>
-        permission?: Permission
-        steps?: number
-      }
-    })
-    .meta({
-      ref: "AgentConfig",
-    })
-  export type Agent = z.infer<typeof Agent>
-
-  export const ExternalAgentConfig = z
-    .object({
-      disabled: z.boolean().optional().describe("Disable this external agent"),
-      path: z.string().optional().describe("Override path to the external agent binary"),
-      model: z.string().optional().describe("Default model for this external agent"),
-      auto_discover: z.boolean().optional().describe("Whether to auto-discover this agent on startup (default: true)"),
-    })
-    .catchall(z.unknown())
-    .meta({
-      ref: "ExternalAgentConfig",
-    })
-  export type ExternalAgentConfig = z.infer<typeof ExternalAgentConfig>
-
-  export const Keybinds = z
-    .object({
-      leader: z.string().optional().default("ctrl+x").describe("Leader key for keybind combinations"),
-      app_exit: z.string().optional().default("ctrl+c,ctrl+d,<leader>q").describe("Exit the application"),
-      editor_open: z.string().optional().default("<leader>e").describe("Open external editor"),
-      theme_list: z.string().optional().default("<leader>t").describe("List available themes"),
-      sidebar_toggle: z.string().optional().default("<leader>b").describe("Toggle sidebar"),
-      scrollbar_toggle: z.string().optional().default("none").describe("Toggle session scrollbar"),
-      username_toggle: z.string().optional().default("none").describe("Toggle username visibility"),
-      status_view: z.string().optional().default("<leader>s").describe("View status"),
-      session_export: z.string().optional().default("<leader>x").describe("Export session to editor"),
-      session_new: z.string().optional().default("<leader>n").describe("Create a new session"),
-      session_list: z.string().optional().default("<leader>l").describe("List all sessions"),
-      session_timeline: z.string().optional().default("<leader>g").describe("Show session timeline"),
-      session_fork: z.string().optional().default("none").describe("Fork session from message"),
-      session_rename: z.string().optional().default("none").describe("Rename session"),
-      session_interrupt: z.string().optional().default("escape").describe("Interrupt current session"),
-      session_compact: z.string().optional().default("<leader>c").describe("Compact the session"),
-      messages_page_up: z.string().optional().default("pageup").describe("Scroll messages up by one page"),
-      messages_page_down: z.string().optional().default("pagedown").describe("Scroll messages down by one page"),
-      messages_half_page_up: z.string().optional().default("ctrl+alt+u").describe("Scroll messages up by half page"),
-      messages_half_page_down: z
-        .string()
-        .optional()
-        .default("ctrl+alt+d")
-        .describe("Scroll messages down by half page"),
-      messages_first: z.string().optional().default("ctrl+g,home").describe("Navigate to first message"),
-      messages_last: z.string().optional().default("ctrl+alt+g,end").describe("Navigate to last message"),
-      messages_next: z.string().optional().default("none").describe("Navigate to next message"),
-      messages_previous: z.string().optional().default("none").describe("Navigate to previous message"),
-      messages_last_user: z.string().optional().default("none").describe("Navigate to last user message"),
-      messages_copy: z.string().optional().default("<leader>y").describe("Copy message"),
-      messages_undo: z.string().optional().default("<leader>u").describe("Undo message"),
-      messages_redo: z.string().optional().default("<leader>r").describe("Redo message"),
-      messages_toggle_conceal: z
-        .string()
-        .optional()
-        .default("<leader>h")
-        .describe("Toggle code block concealment in messages"),
-      tool_details: z.string().optional().default("none").describe("Toggle tool details visibility"),
-      model_list: z.string().optional().default("<leader>m").describe("List available models"),
-      model_cycle_recent: z.string().optional().default("f2").describe("Next recently used model"),
-      model_cycle_recent_reverse: z.string().optional().default("shift+f2").describe("Previous recently used model"),
-      model_cycle_favorite: z.string().optional().default("none").describe("Next favorite model"),
-      model_cycle_favorite_reverse: z.string().optional().default("none").describe("Previous favorite model"),
-      command_list: z.string().optional().default("ctrl+p").describe("List available commands"),
-      agent_list: z.string().optional().default("<leader>a").describe("List agents"),
-      agent_cycle: z.string().optional().default("tab").describe("Next agent"),
-      agent_cycle_reverse: z.string().optional().default("shift+tab").describe("Previous agent"),
-      variant_cycle: z.string().optional().default("ctrl+t").describe("Cycle model variants"),
-      input_clear: z.string().optional().default("ctrl+c").describe("Clear input field"),
-      input_paste: z.string().optional().default("ctrl+v").describe("Paste from clipboard"),
-      input_submit: z.string().optional().default("return").describe("Submit input"),
-      input_newline: z
-        .string()
-        .optional()
-        .default("shift+return,ctrl+return,alt+return,ctrl+j")
-        .describe("Insert newline in input"),
-      input_move_left: z.string().optional().default("left,ctrl+b").describe("Move cursor left in input"),
-      input_move_right: z.string().optional().default("right,ctrl+f").describe("Move cursor right in input"),
-      input_move_up: z.string().optional().default("up").describe("Move cursor up in input"),
-      input_move_down: z.string().optional().default("down").describe("Move cursor down in input"),
-      input_select_left: z.string().optional().default("shift+left").describe("Select left in input"),
-      input_select_right: z.string().optional().default("shift+right").describe("Select right in input"),
-      input_select_up: z.string().optional().default("shift+up").describe("Select up in input"),
-      input_select_down: z.string().optional().default("shift+down").describe("Select down in input"),
-      input_line_home: z.string().optional().default("ctrl+a").describe("Move to start of line in input"),
-      input_line_end: z.string().optional().default("ctrl+e").describe("Move to end of line in input"),
-      input_select_line_home: z
-        .string()
-        .optional()
-        .default("ctrl+shift+a")
-        .describe("Select to start of line in input"),
-      input_select_line_end: z.string().optional().default("ctrl+shift+e").describe("Select to end of line in input"),
-      input_visual_line_home: z.string().optional().default("alt+a").describe("Move to start of visual line in input"),
-      input_visual_line_end: z.string().optional().default("alt+e").describe("Move to end of visual line in input"),
-      input_select_visual_line_home: z
-        .string()
-        .optional()
-        .default("alt+shift+a")
-        .describe("Select to start of visual line in input"),
-      input_select_visual_line_end: z
-        .string()
-        .optional()
-        .default("alt+shift+e")
-        .describe("Select to end of visual line in input"),
-      input_buffer_home: z.string().optional().default("home").describe("Move to start of buffer in input"),
-      input_buffer_end: z.string().optional().default("end").describe("Move to end of buffer in input"),
-      input_select_buffer_home: z
-        .string()
-        .optional()
-        .default("shift+home")
-        .describe("Select to start of buffer in input"),
-      input_select_buffer_end: z.string().optional().default("shift+end").describe("Select to end of buffer in input"),
-      input_delete_line: z.string().optional().default("ctrl+shift+d").describe("Delete line in input"),
-      input_delete_to_line_end: z.string().optional().default("ctrl+k").describe("Delete to end of line in input"),
-      input_delete_to_line_start: z.string().optional().default("ctrl+u").describe("Delete to start of line in input"),
-      input_backspace: z.string().optional().default("backspace,shift+backspace").describe("Backspace in input"),
-      input_delete: z.string().optional().default("ctrl+d,delete,shift+delete").describe("Delete character in input"),
-      input_undo: z.string().optional().default("ctrl+-,super+z").describe("Undo in input"),
-      input_redo: z.string().optional().default("ctrl+.,super+shift+z").describe("Redo in input"),
-      input_word_forward: z
-        .string()
-        .optional()
-        .default("alt+f,alt+right,ctrl+right")
-        .describe("Move word forward in input"),
-      input_word_backward: z
-        .string()
-        .optional()
-        .default("alt+b,alt+left,ctrl+left")
-        .describe("Move word backward in input"),
-      input_select_word_forward: z
-        .string()
-        .optional()
-        .default("alt+shift+f,alt+shift+right")
-        .describe("Select word forward in input"),
-      input_select_word_backward: z
-        .string()
-        .optional()
-        .default("alt+shift+b,alt+shift+left")
-        .describe("Select word backward in input"),
-      input_delete_word_forward: z
-        .string()
-        .optional()
-        .default("alt+d,alt+delete,ctrl+delete")
-        .describe("Delete word forward in input"),
-      input_delete_word_backward: z
-        .string()
-        .optional()
-        .default("ctrl+w,ctrl+backspace,alt+backspace")
-        .describe("Delete word backward in input"),
-      history_previous: z.string().optional().default("up").describe("Previous history item"),
-      history_next: z.string().optional().default("down").describe("Next history item"),
-      session_child_cycle: z.string().optional().default("<leader>right").describe("Next child session"),
-      session_child_cycle_reverse: z.string().optional().default("<leader>left").describe("Previous child session"),
-      session_parent: z.string().optional().default("<leader>up").describe("Go to parent session"),
-      terminal_suspend: z.string().optional().default("ctrl+z").describe("Suspend terminal"),
-      terminal_title_toggle: z.string().optional().default("none").describe("Toggle terminal title"),
-      tips_toggle: z.string().optional().default("<leader>h").describe("Toggle tips on home screen"),
-    })
-    .strict()
-    .meta({
-      ref: "KeybindsConfig",
-    })
-
-  export const Server = z
-    .object({
-      port: z.number().int().positive().optional().describe("Port to listen on"),
-      hostname: z.string().optional().describe("Hostname to listen on"),
-      mdns: z.boolean().optional().describe("Enable mDNS service discovery"),
-      cors: z.array(z.string()).optional().describe("Additional domains to allow for CORS"),
-    })
-    .strict()
-    .meta({
-      ref: "ServerConfig",
-    })
-
-  export const CategoryConfig = z
-    .object({
-      model: z.string().optional().describe("Model to use for this category (e.g., 'sii-openai/GPT-5.2')"),
-      temperature: z.number().optional().describe("Temperature override for this category"),
-      promptAppend: z.string().optional().describe("Additional prompt context to inject for this category"),
-      description: z.string().optional().describe("Description of when to use this category"),
-    })
-    .strict()
-    .meta({
-      ref: "CategoryConfig",
-    })
-  export type CategoryConfig = z.infer<typeof CategoryConfig>
-
-  export const Layout = z.enum(["auto", "stretch"]).meta({
-    ref: "LayoutConfig",
-  })
-  export type Layout = z.infer<typeof Layout>
-
-  export const Learning = z
-    .object({
-      alpha: z.number().min(0).max(1).optional().describe("Q-learning step size / learning rate (default: 0.3)"),
-      qInit: z.number().optional().describe("Optimistic Q-value initialization per reward dimension (default: 1.0)"),
-      dedupIntentThreshold: z
-        .number()
-        .min(0)
-        .max(1)
-        .optional()
-        .describe("Intent cosine similarity threshold for deduplicating experiences (default: 0.85)"),
-      dedupScriptThreshold: z
-        .number()
-        .min(0)
-        .max(1)
-        .optional()
-        .describe("Script cosine similarity threshold for deduplicating experiences (default: 0.8)"),
-      qHistorySize: z
-        .number()
-        .int()
-        .min(1)
-        .optional()
-        .describe("Maximum Q-value history entries per experience (default: 50)"),
-      snapThreshold: z
-        .number()
-        .min(0)
-        .max(1)
-        .optional()
-        .describe("Threshold for snapping reward dimensions to discrete {-1, 0, 1} (default: 0.5)"),
-      legacyRewardConfidence: z
-        .number()
-        .min(0)
-        .max(1)
-        .optional()
-        .describe("Default confidence for legacy scalar reward format (default: 0.3)"),
-      encoderRetries: z
-        .number()
-        .int()
-        .min(0)
-        .optional()
-        .describe("LLM retry count for intent/script/reward generation (default: 3)"),
-      digestToolOutputBudget: z
-        .number()
-        .int()
-        .min(0)
-        .optional()
-        .describe("Max estimated tokens for tool output in turn digest (default: 800)"),
-      encoderToolFieldBudget: z
-        .number()
-        .int()
-        .min(0)
-        .optional()
-        .describe("Max chars per tool input field in encoder context (default: 500)"),
-      encoderToolOutputBudget: z
-        .number()
-        .int()
-        .min(0)
-        .optional()
-        .describe("Max chars for tool output in encoder context (default: 300)"),
-      rewardWeights: z
-        .object({
-          outcome: z.number().optional().describe("Weight for outcome dimension (default: 0.35)"),
-          intent: z.number().optional().describe("Weight for intent dimension (default: 0.25)"),
-          execution: z.number().optional().describe("Weight for execution dimension (default: 0.2)"),
-          orchestration: z.number().optional().describe("Weight for orchestration dimension (default: 0.1)"),
-          expression: z.number().optional().describe("Weight for expression dimension (default: 0.1)"),
-        })
-        .strict()
-        .optional()
-        .describe(
-          "Weights for multi-dimensional reward composition (default: outcome=0.35, intent=0.25, execution=0.2, orchestration=0.1, expression=0.1)",
-        ),
-      rewardDelay: z
-        .number()
-        .int()
-        .min(0)
-        .optional()
-        .describe("Number of subsequent turns to wait before evaluating reward (default: 2)"),
-    })
-    .strict()
-    .meta({ ref: "LearningConfig" })
-  export type Learning = z.infer<typeof Learning>
-
-  export const PassiveRetrieval = z
-    .object({
-      simThreshold: z
-        .number()
-        .min(0)
-        .max(1)
-        .optional()
-        .describe("Minimum cosine similarity for retrieval candidates (default: 0.7)"),
-      topK: z.number().int().min(1).optional().describe("Number of experiences to retrieve (default: 8)"),
-      epsilon: z.number().min(0).max(1).optional().describe("ε-greedy exploration probability (default: 0.1)"),
-      wSim: z.number().min(0).max(1).optional().describe("Weight for similarity in hybrid score (default: 0.5)"),
-      wQ: z.number().min(0).max(1).optional().describe("Weight for Q-value in hybrid score (default: 0.5)"),
-      explorationConstant: z
-        .number()
-        .min(0)
-        .optional()
-        .describe("UCB1 exploration constant — scales √(ln(N)/n) visit-decay bonus (default: 0.5)"),
-    })
-    .strict()
-    .meta({ ref: "PassiveRetrievalConfig" })
-  export type PassiveRetrieval = z.infer<typeof PassiveRetrieval>
-
-  export const REWARD_WEIGHT_DEFAULTS = {
-    outcome: 0.35,
-    intent: 0.25,
-    execution: 0.2,
-    orchestration: 0.1,
-    expression: 0.1,
-  } as const
-
-  export const LEARNING_DEFAULTS = {
-    alpha: 0.3,
-    qInit: 0.5,
-    dedupIntentThreshold: 0.85,
-    dedupScriptThreshold: 0.8,
-    qHistorySize: 50,
-    snapThreshold: 0.5,
-    legacyRewardConfidence: 0.3,
-    encoderRetries: 3,
-    digestToolOutputBudget: 800,
-    encoderToolFieldBudget: 500,
-    encoderToolOutputBudget: 300,
-    rewardWeights: { ...REWARD_WEIGHT_DEFAULTS },
-    rewardDelay: 5,
-  } as const satisfies Required<Learning>
-
-  export const PASSIVE_RETRIEVAL_DEFAULTS = {
-    simThreshold: 0.7,
-    topK: 8,
-    epsilon: 0.1,
-    wSim: 0.5,
-    wQ: 0.5,
-    explorationConstant: 0.5,
-  } as const satisfies Required<PassiveRetrieval>
-
-  export const EvolutionPassive = z
-    .object({
-      encode: z
-        .boolean()
-        .optional()
-        .describe("Learn from conversations — extract intent, reward, and scripts (default: true)"),
-      retrieve: z
-        .union([z.boolean(), PassiveRetrieval])
-        .optional()
-        .describe("Inject relevant past experiences into new conversations (default: true)"),
-      learning: Learning.optional().describe("Hyperparameters for the experience learning pipeline"),
-    })
-    .strict()
-    .meta({ ref: "EvolutionPassive" })
-
-  export const MEMORY_CATEGORIES = [
-    "user",
-    "self",
-    "relationship",
-    "interaction",
-    "workflow",
-    "coding",
-    "writing",
-    "asset",
-    "insight",
-    "knowledge",
-    "personal",
-    "general",
-  ] as const
-  export type MemoryCategory = (typeof MEMORY_CATEGORIES)[number]
-
-  const CategoryRetrieveConfig = z
-    .object({
-      simThreshold: z.number().optional().describe("Minimum similarity for contextual retrieval"),
-      topK: z.number().optional().describe("Maximum contextual entries to retrieve"),
-    })
-    .strict()
-
-  export const EvolutionActive = z
-    .object({
-      retrieve: z
-        .union([
-          z.boolean(),
-          z
-            .object({
-              simThreshold: z
-                .number()
-                .optional()
-                .describe("Default minimum similarity for auto-injection (default: 0.7)"),
-              topK: z
-                .number()
-                .optional()
-                .describe("Default maximum entries per category to contextually retrieve (default: 3)"),
-              categories: z
-                .record(z.enum(MEMORY_CATEGORIES), CategoryRetrieveConfig)
-                .optional()
-                .describe("Per-category contextual retrieval overrides"),
-            })
-            .strict(),
-        ])
-        .optional()
-        .describe("Auto-inject relevant memories into new conversations (default: true)"),
-      memoryDedupThreshold: z
-        .number()
-        .min(0)
-        .max(1)
-        .optional()
-        .describe("Cosine similarity threshold for blocking duplicate memory writes (default: 0.75)"),
-    })
-    .strict()
-    .meta({ ref: "EvolutionActive" })
-
-  export const EvolutionConfig = z
-    .object({
-      passive: z
-        .union([z.boolean(), EvolutionPassive])
-        .optional()
-        .describe("RL-enhanced passive experience learning (default: true)"),
-      active: z
-        .union([z.boolean(), EvolutionActive])
-        .optional()
-        .describe("Agent-initiated active memory curation via memory tools (default: true)"),
-    })
-    .strict()
-    .meta({ ref: "EvolutionConfig" })
-
-  export const Evolution = z
-    .union([z.boolean(), EvolutionConfig])
-    .optional()
-    .describe("Dual-mode evolution system: passive experience learning + active memory curation (default: true)")
-
-  export const ACTIVE_RETRIEVAL_DEFAULTS = {
-    simThreshold: 0.7,
-    topK: 3,
-  } as const
-
-  export interface CategoryRetrieval {
-    simThreshold: number
-    topK: number
-  }
-
-  export interface ActiveRetrieval {
-    enabled: boolean
-    categories: Record<MemoryCategory, CategoryRetrieval>
-  }
-
-  export interface ResolvedEvolution {
-    encode: boolean
-    retrieve: boolean
-    active: boolean
-    activeRetrieval: ActiveRetrieval
-    passiveRetrieval: Required<PassiveRetrieval>
-    memoryDedupThreshold: number
-    learning: Required<Learning>
-  }
-
-  export function resolveEvolution(raw?: z.infer<typeof Evolution>): ResolvedEvolution {
-    function buildCategoryRetrieval(
-      globalSimThreshold?: number,
-      globalTopK?: number,
-      overrides?: Partial<Record<MemoryCategory, { simThreshold?: number; topK?: number }>>,
-    ): Record<MemoryCategory, CategoryRetrieval> {
-      const simThreshold = globalSimThreshold ?? ACTIVE_RETRIEVAL_DEFAULTS.simThreshold
-      const topK = globalTopK ?? ACTIVE_RETRIEVAL_DEFAULTS.topK
-      const result = {} as Record<MemoryCategory, CategoryRetrieval>
-      for (const category of MEMORY_CATEGORIES) {
-        const override = overrides?.[category]
-        result[category] = {
-          simThreshold: override?.simThreshold ?? simThreshold,
-          topK: override?.topK ?? topK,
-        }
-      }
-      return result
-    }
-
-    const defaultCategories = buildCategoryRetrieval()
-    const defaultActiveRetrieval: ActiveRetrieval = { enabled: true, categories: defaultCategories }
-    const disabledActiveRetrieval: ActiveRetrieval = { enabled: false, categories: defaultCategories }
-
-    const MEMORY_DEDUP_THRESHOLD_DEFAULT = 0.75
-
-    if (raw === false)
-      return {
-        encode: false,
-        retrieve: false,
-        active: false,
-        activeRetrieval: disabledActiveRetrieval,
-        passiveRetrieval: { ...PASSIVE_RETRIEVAL_DEFAULTS },
-        memoryDedupThreshold: MEMORY_DEDUP_THRESHOLD_DEFAULT,
-        learning: { ...LEARNING_DEFAULTS },
-      }
-    if (raw === true || raw === undefined)
-      return {
-        encode: true,
-        retrieve: true,
-        active: true,
-        activeRetrieval: defaultActiveRetrieval,
-        passiveRetrieval: { ...PASSIVE_RETRIEVAL_DEFAULTS },
-        memoryDedupThreshold: MEMORY_DEDUP_THRESHOLD_DEFAULT,
-        learning: { ...LEARNING_DEFAULTS },
-      }
-
-    const passive = raw.passive
-    let encode: boolean
-    let retrieve: boolean
-    let learning: Required<Learning>
-    let passiveRetrieval: Required<PassiveRetrieval>
-    if (passive === false) {
-      encode = false
-      retrieve = false
-      learning = { ...LEARNING_DEFAULTS }
-      passiveRetrieval = { ...PASSIVE_RETRIEVAL_DEFAULTS }
-    } else if (passive === true || passive === undefined) {
-      encode = true
-      retrieve = true
-      learning = { ...LEARNING_DEFAULTS }
-      passiveRetrieval = { ...PASSIVE_RETRIEVAL_DEFAULTS }
-    } else {
-      encode = passive.encode ?? true
-      const rawRetrieve = passive.retrieve
-      if (rawRetrieve === false) {
-        retrieve = false
-        passiveRetrieval = { ...PASSIVE_RETRIEVAL_DEFAULTS }
-      } else if (rawRetrieve === true || rawRetrieve === undefined) {
-        retrieve = true
-        passiveRetrieval = { ...PASSIVE_RETRIEVAL_DEFAULTS }
-      } else {
-        retrieve = true
-        passiveRetrieval = { ...PASSIVE_RETRIEVAL_DEFAULTS, ...rawRetrieve }
-      }
-      const rawLearning = passive.learning
-      learning = {
-        ...LEARNING_DEFAULTS,
-        ...rawLearning,
-        rewardWeights: { ...REWARD_WEIGHT_DEFAULTS, ...rawLearning?.rewardWeights },
-      }
-    }
-
-    const activeRaw = raw.active
-    let active: boolean
-    let activeRetrieval: ActiveRetrieval
-    let memoryDedupThreshold = MEMORY_DEDUP_THRESHOLD_DEFAULT
-    if (activeRaw === false) {
-      active = false
-      activeRetrieval = disabledActiveRetrieval
-    } else if (activeRaw === true || activeRaw === undefined) {
-      active = true
-      activeRetrieval = defaultActiveRetrieval
-    } else {
-      active = true
-      memoryDedupThreshold = activeRaw.memoryDedupThreshold ?? MEMORY_DEDUP_THRESHOLD_DEFAULT
-      const r = activeRaw.retrieve
-      if (r === false) {
-        activeRetrieval = disabledActiveRetrieval
-      } else if (r === true || r === undefined) {
-        activeRetrieval = defaultActiveRetrieval
-      } else {
-        activeRetrieval = {
-          enabled: true,
-          categories: buildCategoryRetrieval(r.simThreshold, r.topK, r.categories),
-        }
-      }
-    }
-
-    return { encode, retrieve, active, activeRetrieval, passiveRetrieval, memoryDedupThreshold, learning }
-  }
-
-  export const Identity = z
-    .object({
-      embedding: z
-        .object({
-          baseURL: z.string().optional().describe("Base URL for the embedding API"),
-          apiKey: z.string().optional().describe("API key for the embedding service"),
-          model: z.string().optional().describe("Embedding model name (e.g., 'Qwen/Qwen3-Embedding-8B')"),
-        })
-        .optional()
-        .describe("Embedding model configuration for memory and retrieval"),
-      rerank: z
-        .object({
-          baseURL: z.string().optional().describe("Base URL for the rerank API"),
-          apiKey: z.string().optional().describe("API key for the rerank service (falls back to embedding.apiKey)"),
-          model: z.string().optional().describe("Rerank model name (e.g., 'Qwen/Qwen3-Reranker-8B')"),
-        })
-        .optional()
-        .describe("Rerank model configuration for memory retrieval refinement"),
-      evolution: Evolution,
-      autonomy: z
-        .boolean()
-        .optional()
-        .describe("Enable autonomous background routines like anima daily wake (default: true)"),
-    })
-    .strict()
-    .meta({
-      ref: "IdentityConfig",
-    })
-  export type Identity = z.infer<typeof Identity>
-
-  export const Provider = ModelsDev.Provider.partial()
-    .extend({
-      whitelist: z.array(z.string()).optional(),
-      blacklist: z.array(z.string()).optional(),
-      models: z
-        .record(
-          z.string(),
-          ModelsDev.Model.partial().extend({
-            variants: z
-              .record(
-                z.string(),
-                z
-                  .object({
-                    disabled: z.boolean().optional().describe("Disable this variant for the model"),
-                  })
-                  .catchall(z.any()),
-              )
-              .optional()
-              .describe("Variant-specific configuration"),
-          }),
-        )
-        .optional(),
-      options: z
-        .object({
-          apiKey: z.string().optional(),
-          baseURL: z.string().optional(),
-          enterpriseUrl: z.string().optional().describe("GitHub Enterprise URL for copilot authentication"),
-          setCacheKey: z.boolean().optional().describe("Enable promptCacheKey for this provider (default false)"),
-          timeout: z
-            .union([
-              z
-                .number()
-                .int()
-                .positive()
-                .describe(
-                  "Timeout in milliseconds for requests to this provider. Default is 900000 (15 minutes). Set to false to disable timeout.",
-                ),
-              z.literal(false).describe("Disable timeout for this provider entirely."),
-            ])
-            .optional()
-            .describe(
-              "Timeout in milliseconds for requests to this provider. Default is 900000 (15 minutes). Set to false to disable timeout.",
-            ),
-        })
-        .catchall(z.any())
-        .optional(),
-    })
-    .strict()
-    .meta({
-      ref: "ProviderConfig",
-    })
-  export type Provider = z.infer<typeof Provider>
-
-  export const Info = z
-    .object({
-      $schema: z.string().optional().describe("JSON schema reference for configuration validation"),
-      theme: z.string().optional().describe("Theme name to use for the interface"),
-      keybinds: Keybinds.optional().describe("Custom keybind configurations"),
-      logLevel: Log.Level.optional().describe("Log level"),
-      server: Server.optional().describe("Server configuration for synergy serve and web commands"),
-      command: z.record(z.string(), Command).optional().describe("Command configuration"),
-      timeout: z
-        .object({
-          invoke_sec: z
-            .number()
-            .positive()
-            .optional()
-            .describe("Max wall-clock seconds for one agent turn (default: 900 = 15min)"),
-          provider: z
-            .object({
-              ttfb_sec: z
-                .number()
-                .positive()
-                .optional()
-                .describe(
-                  "Max seconds to wait for first byte (TTFB) from provider. " +
-                    "Accommodates reasoning/thinking models (e.g. o1-pro, deepseek-r1). " +
-                    "Default: 600 = 10min",
-                ),
-              idle_sec: z
-                .number()
-                .min(0)
-                .optional()
-                .describe("Idle timeout in seconds (0 = disable, default: 180 = 3min). Resets on each data chunk."),
-              wall_sec: z
-                .number()
-                .min(0)
-                .optional()
-                .describe(
-                  "Hard wall-clock timeout per HTTP request in seconds " +
-                    "(0 = disabled, default: 0). CAUTION: conflicts with streaming — " +
-                    "will interrupt normal token output. Only enable if you need a " +
-                    "hard cap beyond idle+TTFB",
-                ),
-            })
-            .optional(),
-          tool: z
-            .object({
-              default_sec: z
-                .number()
-                .positive()
-                .optional()
-                .describe("Default timeout per tool execution in seconds (default: 300 = 5min)"),
-              overrides: z
-                .record(z.string(), z.number().positive())
-                .optional()
-                .describe("Per-tool timeout overrides by tool name, e.g. { bash: 600, webfetch: 120 }"),
-            })
-            .optional(),
-        })
-        .optional()
-        .describe("Timeout configuration for agent turns, provider requests, and tool execution"),
-      watcher: z
-        .object({
-          ignore: z.array(z.string()).optional(),
-        })
-        .optional(),
-      plugin: z.string().array().optional(),
-      snapshot: z.boolean().optional(),
-      autoupdate: z
-        .union([z.boolean(), z.literal("notify")])
-        .optional()
-        .describe(
-          "Automatically update to the latest version. Set to true to auto-update, false to disable, or 'notify' to show update notifications",
-        ),
-      disabled_providers: z.array(z.string()).optional().describe("Disable providers that are loaded automatically"),
-      enabled_providers: z
-        .array(z.string())
-        .optional()
-        .describe("When set, ONLY these providers will be enabled. All other providers will be ignored"),
-      model: z
-        .string()
-        .describe("Default model in the format of provider/model, eg anthropic/claude-sonnet-4-5")
-        .optional(),
-      nano_model: z
-        .string()
-        .describe(
-          "Cheapest model for trivial extraction tasks like title generation, in the format of provider/model. Falls back to mini_model → mid_model → model.",
-        )
-        .optional(),
-      mini_model: z
-        .string()
-        .describe(
-          "Lightweight model for simple tasks like intent extraction, in the format of provider/model. Falls back to mid_model → model.",
-        )
-        .optional(),
-      mid_model: z
-        .string()
-        .describe(
-          "Mid-tier model for internal agents that need moderate reasoning (script extraction, reward evaluation, code exploration), in the format of provider/model. Falls back to the default model.",
-        )
-        .optional(),
-      thinking_model: z
-        .string()
-        .describe(
-          "Deep thinking model for complex reasoning and architecture tasks, in the format of provider/model. Falls back to the default model if not set.",
-        )
-        .optional(),
-      long_context_model: z
-        .string()
-        .describe(
-          "Model with extra-large context window for processing very long inputs, in the format of provider/model. Falls back to the default model if not set.",
-        )
-        .optional(),
-      creative_model: z
-        .string()
-        .describe(
-          "Model for creative and visual tasks (UI design, writing, artistry), in the format of provider/model. Falls back to the default model if not set.",
-        )
-        .optional(),
-      holos_friend_reply_model: z
-        .string()
-        .describe(
-          "Model for Holos automatic friend replies, in the format of provider/model. Falls back to the default model if not set.",
-        )
-        .optional(),
-      vision_model: z
-        .string()
-        .describe(
-          "Model for image analysis via the look_at tool, in the format of provider/model. If not set, look_at is disabled.",
-        )
-        .optional(),
-      default_agent: z
-        .string()
-        .optional()
-        .describe(
-          "Default agent to use when none is specified. Must be a primary agent. Falls back to 'synergy' if not set or if the specified agent is invalid.",
-        ),
-      username: z
-        .string()
-        .optional()
-        .describe("Custom username to display in conversations instead of system username"),
-      agent: z
-        .object({
-          // primary
-          synergy: Agent.optional(),
-          "synergy-max": Agent.optional(),
-          // classic subagents
-          developer: Agent.optional(),
-          // subagent
-          general: Agent.optional(),
-          explore: Agent.optional(),
-          // specialized
-          title: Agent.optional(),
-          summary: Agent.optional(),
-          compaction: Agent.optional(),
-        })
-        .catchall(Agent)
-        .optional()
-        .describe("Agent configuration"),
-      external_agent: z
-        .record(z.string(), ExternalAgentConfig)
-        .optional()
-        .describe("External agent configurations (e.g. codex, claude-code)"),
-      provider: z
-        .record(z.string(), Provider)
-        .optional()
-        .describe("Custom provider configurations and model overrides"),
-      identity: Identity.optional().describe("Identity configuration for embedding and evolution"),
-      mcp: z
-        .record(
-          z.string(),
-          z.union([
-            Mcp,
-            z
-              .object({
-                enabled: z.boolean(),
-              })
-              .strict(),
-          ]),
-        )
-        .optional()
-        .describe("MCP (Model Context Protocol) server configurations"),
-      channel: z
-        .record(z.string(), Channel)
-        .optional()
-        .describe("Channel configurations for messaging platform integrations"),
-      holos: Holos.optional().describe("Holos platform configuration"),
-      email: Email.optional().describe("Outgoing email configuration"),
-      formatter: z
-        .union([
-          z.literal(false),
-          z.record(
-            z.string(),
-            z.object({
-              disabled: z.boolean().optional(),
-              command: z.array(z.string()).optional(),
-              environment: z.record(z.string(), z.string()).optional(),
-              extensions: z.array(z.string()).optional(),
-            }),
-          ),
-        ])
-        .optional(),
-      lsp: z
-        .union([
-          z.literal(false),
-          z.record(
-            z.string(),
-            z.union([
-              z.object({
-                disabled: z.literal(true),
-              }),
-              z.object({
-                command: z.array(z.string()),
-                extensions: z.array(z.string()).optional(),
-                disabled: z.boolean().optional(),
-                env: z.record(z.string(), z.string()).optional(),
-                initialization: z.record(z.string(), z.any()).optional(),
-              }),
-            ]),
-          ),
-        ])
-        .optional()
-        .refine(
-          (data) => {
-            if (!data) return true
-            if (typeof data === "boolean") return true
-            const serverIds = new Set(Object.values(LSPServer).map((s) => s.id))
-
-            return Object.entries(data).every(([id, config]) => {
-              if (config.disabled) return true
-              if (serverIds.has(id)) return true
-              return Boolean(config.extensions)
-            })
-          },
-          {
-            error: "For custom LSP servers, 'extensions' array is required.",
-          },
-        ),
-      instructions: z.array(z.string()).optional().describe("Additional instruction files or patterns to include"),
-      layout: Layout.optional().describe("@deprecated Always uses stretch layout."),
-      permission: Permission.optional(),
-      tools: z.record(z.string(), z.boolean()).optional(),
-      enterprise: z
-        .object({
-          url: z.string().optional().describe("Enterprise URL"),
-        })
-        .optional(),
-      agora: z
-        .object({
-          url: z.string().optional().describe("Agora API base URL (defaults to https://agora.holosai.io)"),
-          tokenUrl: z
-            .string()
-            .optional()
-            .describe("Holos API URL for Agora token exchange (defaults to https://www.holosai.io)"),
-          giteaSSHHost: z.string().optional().describe("Override SSH hostname used for Agora's Gitea remote"),
-        })
-
-        .optional()
-        .describe("Agora Q&A platform configuration"),
-      question: z
-        .object({
-          timeout: z
-            .number()
-            .min(0)
-            .optional()
-            .describe("Seconds before unanswered questions auto-expire (0 = no timeout, default 1800 = 30min)"),
-        })
-        .optional(),
-      compaction: z
-        .object({
-          auto: z.boolean().optional().describe("Enable automatic compaction when context is full (default: true)"),
-          prune: z.boolean().optional().describe("Enable pruning of old tool outputs (default: true)"),
-          overflowThreshold: z
-            .number()
-            .min(0.5)
-            .max(1)
-            .optional()
-            .describe("Fraction of usable context that triggers auto-compaction (default: 0.85)"),
-          maxHistoryImages: z
-            .number()
-            .int()
-            .optional()
-            .describe(
-              "Maximum number of historical images to send as base64 per request (older images replaced with text placeholders). Default: 8.",
-            ),
-        })
-        .optional(),
-      experimental: z
-        .object({
-          batch_tool: z.boolean().optional().describe("Enable the batch tool"),
-          openTelemetry: z
-            .boolean()
-            .optional()
-            .describe("Enable OpenTelemetry spans for AI SDK calls (using the 'experimental_telemetry' flag)"),
-          primary_tools: z
-            .array(z.string())
-            .optional()
-            .describe("Tools that should only be available to primary agents."),
-          continue_loop_on_deny: z.boolean().optional().describe("Continue the agent loop when a tool call is denied"),
-          mcp_timeout: z
-            .number()
-            .int()
-            .positive()
-            .optional()
-            .describe("Timeout in milliseconds for model context protocol (MCP) requests"),
-        })
-        .optional(),
-      pluginConfig: z
-        .record(z.string(), z.record(z.string(), z.any()))
-        .optional()
-        .describe("Per-plugin configuration namespaces. Keys are plugin IDs, values are plugin-specific config."),
-      category: z
-        .record(z.string(), CategoryConfig)
-        .optional()
-        .describe("Custom category configurations for background tasks. Categories define model and prompt presets."),
-    })
-    .strict()
-    .meta({
-      ref: "Config",
-    })
-
-  export type Info = z.output<typeof Info>
 
   export const global = lazy(async () => {
     const activeSet = await ConfigSet.activeName()
@@ -1895,8 +671,8 @@ export namespace Config {
         if (account?.appSecret) account.appSecret = REDACTED_SENTINEL
       }
     }
-    if (result.identity?.embedding?.apiKey) result.identity.embedding.apiKey = REDACTED_SENTINEL
-    if (result.identity?.rerank?.apiKey) result.identity.rerank.apiKey = REDACTED_SENTINEL
+    if (result.embedding?.apiKey) result.embedding.apiKey = REDACTED_SENTINEL
+    if (result.rerank?.apiKey) result.rerank.apiKey = REDACTED_SENTINEL
     if (result.provider) {
       for (const provider of Object.values(result.provider) as any[]) {
         if (provider?.options?.apiKey) provider.options.apiKey = REDACTED_SENTINEL
@@ -1931,11 +707,11 @@ export namespace Config {
         }
       }
     }
-    if (result.identity?.embedding?.apiKey === REDACTED_SENTINEL && stored.identity?.embedding?.apiKey) {
-      result.identity.embedding.apiKey = stored.identity.embedding.apiKey
+    if (result.embedding?.apiKey === REDACTED_SENTINEL && stored.embedding?.apiKey) {
+      result.embedding.apiKey = stored.embedding.apiKey
     }
-    if (result.identity?.rerank?.apiKey === REDACTED_SENTINEL && stored.identity?.rerank?.apiKey) {
-      result.identity.rerank.apiKey = stored.identity.rerank.apiKey
+    if (result.rerank?.apiKey === REDACTED_SENTINEL && stored.rerank?.apiKey) {
+      result.rerank.apiKey = stored.rerank.apiKey
     }
     if (result.provider && stored.provider) {
       for (const [key, provider] of Object.entries(result.provider) as [string, any][]) {
