+++ packages/synergy/src/config/setup.ts
@@ -731,28 +731,20 @@ export namespace ConfigSetup {
     }
   }
 
-  function extractIdentityInputs(config: SetupDraft | Config.Info) {
+  function extractEmbeddingRerank(config: SetupDraft | Config.Info) {
     const draft = config as SetupDraft
     const info = config as Config.Info
 
     const embedding =
       draft.embedding ??
-      (info.identity?.embedding?.baseURL && info.identity.embedding.apiKey && info.identity.embedding.model
-        ? {
-            baseURL: info.identity.embedding.baseURL,
-            apiKey: info.identity.embedding.apiKey,
-            model: info.identity.embedding.model,
-          }
+      (info.embedding?.baseURL && info.embedding.apiKey && info.embedding.model
+        ? { baseURL: info.embedding.baseURL, apiKey: info.embedding.apiKey, model: info.embedding.model }
         : undefined)
 
     const rerank =
       draft.rerank ??
-      (info.identity?.rerank?.baseURL && info.identity.rerank.apiKey && info.identity.rerank.model
-        ? {
-            baseURL: info.identity.rerank.baseURL,
-            apiKey: info.identity.rerank.apiKey,
-            model: info.identity.rerank.model,
-          }
+      (info.rerank?.baseURL && info.rerank.apiKey && info.rerank.model
+        ? { baseURL: info.rerank.baseURL, apiKey: info.rerank.apiKey, model: info.rerank.model }
         : undefined)
 
     return { embedding, rerank }
@@ -1241,20 +1233,20 @@ export namespace ConfigSetup {
     }
   }
 
-  function stripFailedIdentityFields(
-    identity: Record<string, unknown>,
+  function disableEngramForMissingEmbedding(
+    next: { engram?: Record<string, unknown> },
     validation: RequiredCoreValidationResult,
   ): void {
     const embeddingFailed = validation.fields.embedding.failedRecommended
     const rerankFailed = validation.fields.rerank.failedRecommended
     if (!embeddingFailed && !rerankFailed) return
-    if (embeddingFailed) delete identity.embedding
-    if (rerankFailed) delete identity.rerank
-    identity.evolution = false
+    if (!next.engram) next.engram = {}
+    next.engram.memory = { enabled: false }
+    next.engram.experience = { encode: false, retrieve: false }
   }
 
   export async function validateRequiredCore(config: SetupDraft): Promise<RequiredCoreValidationResult> {
-    const { embedding, rerank } = extractIdentityInputs(config)
+    const { embedding, rerank } = extractEmbeddingRerank(config)
     const context: ImportProbeContext = { config: applySetupDraft({}, config), auth: config.auth }
     const fields: Record<RequiredCoreField, FieldValidationResult> = {
       model: config.model
@@ -1267,16 +1259,12 @@ export namespace ConfigSetup {
             importContext: context,
           })
         : requiredMissing("Vision model is required"),
-      embedding: downgradeRecommended(
-        embedding
-          ? await verifyIdentityModel(embedding, "embedding")
-          : recommendedSkipped("Skipped — memory evolution will be disabled"),
-      ),
-      rerank: downgradeRecommended(
-        rerank
-          ? await verifyIdentityModel(rerank, "rerank")
-          : recommendedSkipped("Skipped — memory evolution will be disabled"),
-      ),
+      embedding: embedding
+        ? await verifyIdentityModel(embedding, "embedding")
+        : recommendedSkipped("No embedding configured — using local model"),
+      rerank: rerank
+        ? downgradeRecommended(await verifyIdentityModel(rerank, "rerank"))
+        : recommendedSkipped("No rerank configured — reranking disabled"),
     }
 
     return summarizeValidation(fields)
@@ -1286,7 +1274,7 @@ export namespace ConfigSetup {
     config: SetupDraft,
     options: LiveProbeOptions = {},
   ): Promise<RequiredCoreValidationResult> {
-    const { embedding, rerank } = extractIdentityInputs(config)
+    const { embedding, rerank } = extractEmbeddingRerank(config)
     const context: ImportProbeContext = { config: applySetupDraft({}, config), auth: config.auth }
     const [model, visionModel, embeddingResult, rerankResult] = await Promise.all([
       config.model
@@ -1301,7 +1289,7 @@ export namespace ConfigSetup {
         : Promise.resolve(requiredMissing("Vision model is required")),
       embedding
         ? probeEmbeddingModel(embedding)
-        : Promise.resolve(recommendedSkipped("Skipped — memory evolution will be disabled")),
+        : Promise.resolve(recommendedSkipped("No embedding configured — using local model")),
       rerank
         ? options.rerank === "skip"
           ? Promise.resolve({
@@ -1310,7 +1298,7 @@ export namespace ConfigSetup {
               message: "Rerank live probe skipped",
             })
           : probeRerankModel(rerank)
-        : Promise.resolve(recommendedSkipped("Skipped — memory evolution will be disabled")),
+        : Promise.resolve(recommendedSkipped("No rerank configured — reranking disabled")),
     ])
 
     return summarizeValidation({
@@ -1322,7 +1310,7 @@ export namespace ConfigSetup {
   }
 
   async function validateImportedCore(config: Config.Info): Promise<RequiredCoreValidationResult> {
-    const { embedding, rerank } = extractIdentityInputs(config)
+    const { embedding, rerank } = extractEmbeddingRerank(config)
     const [model, visionModel, embeddingResult, rerankResult] = await Promise.all([
       config.model
         ? verifyLanguageModel(config.model, { label: "Default model", importContext: { config } })
@@ -1336,10 +1324,10 @@ export namespace ConfigSetup {
         : Promise.resolve(requiredMissing("Vision model is required")),
       embedding
         ? verifyIdentityModel(embedding, "embedding")
-        : Promise.resolve(recommendedSkipped("Skipped — memory evolution will be disabled")),
+        : Promise.resolve(recommendedSkipped("No embedding configured — using local model")),
       rerank
         ? verifyIdentityModel(rerank, "rerank")
-        : Promise.resolve(recommendedSkipped("Skipped — memory evolution will be disabled")),
+        : Promise.resolve(recommendedSkipped("No rerank configured — reranking disabled")),
     ])
 
     return summarizeValidation({
@@ -1354,7 +1342,7 @@ export namespace ConfigSetup {
     config: Config.Info,
     options: LiveProbeOptions = {},
   ): Promise<RequiredCoreValidationResult> {
-    const { embedding, rerank } = extractIdentityInputs(config)
+    const { embedding, rerank } = extractEmbeddingRerank(config)
     const [model, visionModel, embeddingResult, rerankResult] = await Promise.all([
       config.model
         ? probeLanguageModel(config.model, { label: "Default model", importContext: { config } })
@@ -1368,7 +1356,7 @@ export namespace ConfigSetup {
         : Promise.resolve(requiredMissing("Vision model is required")),
       embedding
         ? probeEmbeddingModel(embedding)
-        : Promise.resolve(recommendedSkipped("Skipped — memory evolution will be disabled")),
+        : Promise.resolve(recommendedSkipped("No embedding configured — using local model")),
       rerank
         ? options.rerank === "skip"
           ? Promise.resolve({
@@ -1377,7 +1365,7 @@ export namespace ConfigSetup {
               message: "Rerank live probe skipped",
             })
           : probeRerankModel(rerank)
-        : Promise.resolve(recommendedSkipped("Skipped — memory evolution will be disabled")),
+        : Promise.resolve(recommendedSkipped("No rerank configured — reranking disabled")),
     ])
 
     return summarizeValidation({
@@ -1400,35 +1388,33 @@ export namespace ConfigSetup {
       else delete next[key]
     }
 
-    const identity = { ...(base.identity ?? {}) }
     const hasEmbedding = Boolean(draft.embedding?.baseURL && draft.embedding?.apiKey && draft.embedding?.model)
     const hasRerank = Boolean(draft.rerank?.baseURL && draft.rerank?.apiKey && draft.rerank?.model)
 
+    // Embedding — always available (local fallback), but preserve user's explicit remote config
     if (hasEmbedding) {
-      identity.embedding = {
+      next.embedding = {
         baseURL: draft.embedding!.baseURL,
         apiKey: draft.embedding!.apiKey,
         model: draft.embedding!.model,
       }
     } else {
-      delete identity.embedding
+      delete next.embedding
     }
 
+    // Rerank
     if (hasRerank) {
-      identity.rerank = {
+      next.rerank = {
         baseURL: draft.rerank!.baseURL,
         apiKey: draft.rerank!.apiKey,
         model: draft.rerank!.model,
       }
     } else {
-      delete identity.rerank
+      delete next.rerank
     }
 
-    identity.evolution = hasEmbedding && hasRerank
-
-    if (Object.keys(identity).length > 0) next.identity = identity
-    else delete next.identity
-
+    // Engram — always enabled (local embedding always works)
+    next.engram = { memory: { enabled: true }, experience: { encode: true, retrieve: true }, autonomy: true }
     if (draft.provider) {
       const provider = { ...(base.provider ?? {}) }
       for (const [providerID, config] of Object.entries(draft.provider)) {
@@ -1470,10 +1456,7 @@ export namespace ConfigSetup {
     const current = await Config.globalRaw()
     const built = applySetupDraft(current, config)
 
-    if (built.identity) {
-      stripFailedIdentityFields(built.identity, resolvedValidation)
-      if (Object.keys(built.identity).length === 0) delete built.identity
-    }
+    disableEngramForMissingEmbedding(built as { engram?: Record<string, unknown> }, resolvedValidation)
 
     const parsed = Config.Info.safeParse(built)
     if (!parsed.success) {
@@ -1511,8 +1494,9 @@ export namespace ConfigSetup {
     "default_agent",
     "username",
     "agent",
-    "provider",
-    "identity",
+    "embedding",
+    "rerank",
+    "engram",
     "mcp",
     "channel",
     "formatter",
@@ -1636,14 +1620,6 @@ export namespace ConfigSetup {
     const out: Record<string, unknown> = { $schema: CONFIG_SCHEMA }
     for (const [key, value] of Object.entries(result.config)) {
       if (value !== undefined && value !== null) out[key] = value
-    }
-
-    if (result.coreValidation) {
-      const identity = out.identity as Record<string, unknown> | undefined
-      if (identity) {
-        stripFailedIdentityFields(identity, result.coreValidation)
-        if (Object.keys(identity).length === 0) delete out.identity
-      }
     }
 
     const parsed = Config.Info.safeParse(out)
