+++ packages/synergy/src/engram/embedding.ts
@@ -1,6 +1,7 @@
 import z from "zod"
 import { embed } from "ai"
 import { createOpenAICompatible } from "@ai-sdk/openai-compatible"
+import { pipeline } from "@huggingface/transformers"
 import { Log } from "../util/log"
 import { Config } from "../config/config"
 
@@ -10,6 +11,24 @@ export namespace Embedding {
   const DEFAULT_BASE_URL = "https://api.siliconflow.cn/v1"
   const DEFAULT_MODEL = "Qwen/Qwen3-Embedding-8B"
   const TIMEOUT_MS = 10_000
+
+  // Local model singleton — created on first use
+  let localExtractor: any
+  let localModelReady = false
+  let localModelError: Error | undefined
+
+  async function getLocalExtractor() {
+    if (localExtractor) return localExtractor
+    if (localModelError) throw localModelError
+    try {
+      localExtractor = await pipeline("feature-extraction", "Xenova/all-MiniLM-L6-v2", { dtype: "q8" })
+      localModelReady = true
+      return localExtractor
+    } catch (err) {
+      localModelError = err instanceof Error ? err : new Error(String(err))
+      throw localModelError
+    }
+  }
 
   export const Info = z
     .object({
@@ -28,14 +47,23 @@ export namespace Embedding {
   export async function generate(input: Input): Promise<Info> {
     using _ = log.time("generate", { id: input.id })
     const resolved = await resolveModel()
-    const { embedding } = await embed({
-      model: resolved.model,
-      value: input.text,
-      abortSignal: AbortSignal.timeout(TIMEOUT_MS),
-    })
+
+    let vector: number[]
+    if (resolved.mode === "local") {
+      const result = await resolved.extractor(input.text, { pooling: "mean", normalize: true })
+      vector = Array.from(result.data as Float32Array)
+    } else {
+      const { embedding } = await embed({
+        model: resolved.model,
+        value: input.text,
+        abortSignal: AbortSignal.timeout(TIMEOUT_MS),
+      })
+      vector = embedding as number[]
+    }
+
     return {
       id: input.id,
-      vector: embedding,
+      vector,
       model: resolved.modelID,
     }
   }
@@ -45,27 +73,52 @@ export namespace Embedding {
     return Promise.all(inputs.map(generate))
   }
 
+  /**
+   * Start async warmup of the local embedding model.
+   * Call this on server startup. Does not block — failures are silent.
+   * While warmup is in progress, embedding calls that need the local model
+   * will block until ready, so callers should use .catch() fallbacks.
+   */
+  export function warmup() {
+    if (localModelReady || localModelError) return
+    getLocalExtractor()
+      .then(() => {
+        log.info("local embedding model ready")
+      })
+      .catch((err) => {
+        log.warn("local embedding model warmup failed", { error: err instanceof Error ? err.message : String(err) })
+      })
+  }
+
   async function resolveModel() {
     const config = await Config.get()
-    const embeddingConfig = config.identity?.embedding
+    const ec = config.embedding
 
-    const baseURL = embeddingConfig?.baseURL ?? DEFAULT_BASE_URL
-    const apiKey = embeddingConfig?.apiKey
-    const modelName = embeddingConfig?.model ?? DEFAULT_MODEL
-
-    if (!apiKey) {
-      throw new Error("Embedding API key is required. Configure it in identity.embedding.apiKey in your synergy.jsonc.")
+    // User explicitly configured remote → use remote
+    if (ec?.apiKey) {
+      const baseURL = ec.baseURL ?? DEFAULT_BASE_URL
+      const modelName = ec.model ?? DEFAULT_MODEL
+      const provider = createOpenAICompatible({
+        name: "embedding",
+        baseURL,
+        apiKey: ec.apiKey,
+      })
+      return {
+        mode: "remote" as const,
+        model: provider.textEmbeddingModel(modelName),
+        modelID: modelName,
+      }
     }
 
-    const provider = createOpenAICompatible({
-      name: "embedding",
-      baseURL,
-      apiKey,
-    })
-
-    return {
-      model: provider.textEmbeddingModel(modelName),
-      modelID: modelName,
+    // Zero-config → local
+    try {
+      const extractor = await getLocalExtractor()
+      return { mode: "local" as const, extractor, modelID: "Xenova/all-MiniLM-L6-v2", dimensions: 384 }
+    } catch (err) {
+      throw new Error(
+        `Embedding unavailable: no API key configured and local model failed to load. ` +
+          `Configure embedding config or check network. ${err instanceof Error ? err.message : String(err)}`,
+      )
     }
   }
 }
