+++ packages/synergy/src/engram/experience-encoder.ts
@@ -63,10 +63,16 @@ export namespace ExperienceEncoder {
     const scope = session.scope as Scope
 
     const config = await Config.get()
-    const evo = Config.resolveEvolution(config.identity?.evolution)
-    if (evo.encode === false) return { encoded: false, skipped: true }
+    const engram = (config as any).engram as
+      | { experience?: { encode?: boolean; learning?: Config.Learning } }
+      | undefined
+    if (engram?.experience?.encode === false) return { encoded: false, skipped: true }
 
-    const learning = evo.learning
+    const learning = {
+      ...Config.LEARNING_DEFAULTS,
+      ...engram?.experience?.learning,
+      rewardWeights: { ...Config.REWARD_WEIGHT_DEFAULTS, ...engram?.experience?.learning?.rewardWeights },
+    } as Required<Config.Learning>
     const msgs = await Session.messages({ sessionID })
 
     userMessageID = Turn.resolveRealUser(msgs, userMessageID)
@@ -236,10 +242,16 @@ export namespace ExperienceEncoder {
     if (SessionEndpoint.isHolos(session?.endpoint)) return
 
     const config = await Config.get()
-    const evo = Config.resolveEvolution(config.identity?.evolution)
-    if (evo.encode === false) return
+    const engram = (config as any).engram as
+      | { experience?: { encode?: boolean; learning?: Config.Learning } }
+      | undefined
+    if (engram?.experience?.encode === false) return
 
-    const learning = evo.learning
+    const learning = {
+      ...Config.LEARNING_DEFAULTS,
+      ...engram?.experience?.learning,
+      rewardWeights: { ...Config.REWARD_WEIGHT_DEFAULTS, ...engram?.experience?.learning?.rewardWeights },
+    } as Required<Config.Learning>
     const pending = EngramDB.Experience.listPendingRewards(sessionID)
     if (pending.length === 0) return
